requirements.txt
PyQt5>=5.15.0
pytesseract>=0.3.0
opencv-python>=4.5.0
numpy>=1.21.0
requests>=2.25.0
keyboard>=0.13.0
pyperclip>=1.8.0
Pillow>=8.0.0